pyproject.toml
[tool.poetry]
name = "slackapp-chatgpt"
version = "0.1.0"
description = ""
authors = ["yuichi.nasukawa <nasu0227@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.0"
langchain = "^0.0.350"
openai = "^1.4.0"
moment = "^0.12.1"
momento = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
